+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+## v1.0.1 (Released 23-01-2024)
+- Randomization fixes
+- Flare now is noisy for enemies (attracts dogs)
+- Possible light and activation fixes
+- Configurable rarity and noise settings
 ## v1.0.0 (Released 08-01-2024)
 - Initial release